requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
bidict==0.23.1
blinker==1.8.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
distro==1.9.0
dnspython==1.16.0
eventlet==0.30.2
fastjsonschema==2.19.1
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Failsafe==0.2
Flask-SocketIO==5.3.6
frozenlist==1.4.1
gevent==24.2.1
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-texttospeech==2.16.3
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.2
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
MarkupSafe==2.1.5
multidict==6.0.5
nbformat==5.10.4
openai==1.25.0
packaging==24.0
pigar==2.1.4
platformdirs==4.2.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
pyparsing==3.1.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
setuptools==69.5.1
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
tqdm==4.66.2
traitlets==5.14.3
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
Wave==0.0.2
Werkzeug==3.0.2
wsproto==1.2.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.3
